requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# AI
openai==1.51.0

# Telegram
python-telegram-bot==21.6

# Scheduler
apscheduler==3.10.4

# HTTP client
httpx==0.27.2

# Web templates
jinja2==3.1.4
python-multipart==0.0.12

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
